updat ed close button to clear state on all views, inc. loading
