+++ src/Components/ApiError.js
@@ -1,9 +1,9 @@
 import { Link } from "react-router-dom"
 
-const ApiError = () => {
+const ApiError = ({ handleClose }) => {
     return (
         <>
-            <Link to="/" className="close">
+            <Link to="/" className="close" onClick={handleClose}>
                 <p className="sr-only">To Home</p>
                 <i className="fa-solid fa-circle-xmark"></i>
             </Link>
