+++ src/Components/Loading.js
@@ -1,11 +1,19 @@
-const Loading = () => {
+import { Link } from "react-router-dom"
+
+const Loading = ({ handleClose }) => {
     return (
-        <div className="loading">
-            <i className="fa-solid fa-tree"></i>
-            <i className="fa-solid fa-tree"></i>
-            <i className="fa-solid fa-tree"></i>
-            <p className="sr-only">Loading</p>
-        </div>
+        <>
+            <Link to="/" className="close" onClick={handleClose}>
+                <p className="sr-only">To Home</p>
+                <i className="fa-solid fa-circle-xmark"></i>
+            </Link>
+            <div className="loading">
+                <i className="fa-solid fa-tree"></i>
+                <i className="fa-solid fa-tree"></i>
+                <i className="fa-solid fa-tree"></i>
+                <p className="sr-only">Loading</p>
+            </div>
+        </>
     );
 };
